move MarklightTextProcessingResult into a new file
